Fix formatting, move FpsCalculator to lib
